requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
black==23.12.1
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   pyproj
    #   requests
cffi==1.16.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
cryptography==41.0.7
    # via
    #   jwcrypto
    #   python-jose
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecated==1.2.14
    # via jwcrypto
distlib==0.3.8
    # via virtualenv
django==4.2.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-oauth-toolkit
    #   djangorestframework
    #   social-auth-app-django
django-cors-headers==4.3.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-model-utils==4.4.0
    # via -r requirements.in
django-oauth-toolkit==2.3.0
    # via drf-social-oauth2
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-social-oauth2
drf-social-oauth2==2.1.3
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
filelock==3.13.1
    # via virtualenv
flake8==7.0.0
    # via -r requirements.in
identify==2.5.35
    # via pre-commit
idna==3.6
    # via requests
isort==5.13.2
    # via -r requirements.in
jwcrypto==1.5.1
    # via django-oauth-toolkit
markdown==3.5.2
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
oauthlib==3.2.2
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via black
pathspec==0.12.1
    # via black
pillow==10.2.0
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   black
    #   virtualenv
pre-commit==3.6.2
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.11.1
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.2.0
    # via flake8
pyjwt==2.8.0
    # via social-auth-core
pyproj==3.6.1
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via drf-social-oauth2
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via pre-commit
requests==2.31.0
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
social-auth-app-django==5.4.0
    # via drf-social-oauth2
social-auth-core==4.5.1
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
urllib3==2.1.0
    # via requests
virtualenv==20.25.1
    # via pre-commit
wrapt==1.16.0
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
